+++ lib/model/entry/extensions/metadata_edition.dart
@@ -32,7 +32,7 @@ extension ExtraAvesEntryMetadataEdition on AvesEntry {
     final appliedModifier = await _applyDateModifierToEntry(userModifier);
     if (appliedModifier == null) {
       if (isValid && userModifier.action != DateEditAction.copyField) {
-        await reportService.recordError('failed to get date for modifier=$userModifier, entry=$this', null);
+        await reportService.recordError('failed to get date for modifier=$userModifier, entry=$this');
       }
       return {};
     }
@@ -65,7 +65,7 @@ extension ExtraAvesEntryMetadataEdition on AvesEntry {
                   final shiftedDate = date.add(Duration(seconds: appliedModifier.shiftSeconds!));
                   editCreateDateXmp(descriptions, shiftedDate);
                 } else {
-                  reportService.recordError('failed to parse XMP date=$xmpDate', null);
+                  reportService.recordError('failed to parse XMP date=$xmpDate');
                 }
               }
             case DateEditAction.remove: