+++ lib/services/device_service.dart
@@ -8,9 +8,9 @@ abstract class DeviceService {
 
   Future<Map<String, dynamic>> getCapabilities();
 
-  Future<int?> getDefaultTimeZoneRawOffsetMillis();
-
   Future<List<Locale>> getLocales();
+
+  Future<void> setLocaleConfig(List<Locale> locales);
 
   Future<int> getPerformanceClass();
 
@@ -51,16 +51,6 @@ class PlatformDeviceService implements DeviceService {
   }
 
   @override
-  Future<int?> getDefaultTimeZoneRawOffsetMillis() async {
-    try {
-      return await _platform.invokeMethod('getDefaultTimeZoneRawOffsetMillis');
-    } on PlatformException catch (e, stack) {
-      await reportService.recordError(e, stack);
-    }
-    return null;
-  }
-
-  @override
   Future<List<Locale>> getLocales() async {
     try {
       final result = await _platform.invokeMethod('getLocales');
@@ -78,6 +68,17 @@ class PlatformDeviceService implements DeviceService {
       await reportService.recordError(e, stack);
     }
     return [];
+  }
+
+  @override
+  Future<void> setLocaleConfig(List<Locale> locales) async {
+    try {
+      await _platform.invokeMethod('setLocaleConfig', <String, dynamic>{
+        'locales': locales.map((v) => v.toLanguageTag()).toList(),
+      });
+    } on PlatformException catch (e, stack) {
+      await reportService.recordError(e, stack);
+    }
   }
 
   @override