+++ lib/widgets/filter_grids/common/filter_grid_page.dart
@@ -286,7 +286,7 @@ class _FilterGridContentState<T extends CollectionFilter> extends State<_FilterG
   Widget build(BuildContext context) {
     final source = context.read<CollectionSource>();
     final settingsRouteKey = context.read<TileExtentController>().settingsRouteKey;
-    final tileLayout = context.select<Settings, TileLayout>((s) => s.getTileLayout(settingsRouteKey));
+    final tileLayout = context.select<Settings, TileLayout>((v) => v.getTileLayout(settingsRouteKey));
     return Selector<Query, bool>(
       selector: (context, query) => query.enabled,
       builder: (context, queryEnabled, child) {