+++ plugins/aves_model/lib/src/settings/keys.dart
@@ -3,7 +3,7 @@ class SettingKeys {
 
   static const Set<String> _internalKeys = {
     hasAcceptedTermsKey,
-    catalogTimeZoneRawOffsetMillisKey,
+    catalogTimeZoneOffsetMillisKey,
     searchHistoryKey,
     platformAccelerometerRotationKey,
     platformTransitionAnimationScaleKey,
@@ -21,7 +21,7 @@ class SettingKeys {
   static const isErrorReportingAllowedKey = 'is_crashlytics_enabled';
   static const localeKey = 'locale';
   static const forceWesternArabicNumeralsKey = 'force_western_arabic_numerals';
-  static const catalogTimeZoneRawOffsetMillisKey = 'catalog_time_zone_raw_offset_millis';
+  static const catalogTimeZoneOffsetMillisKey = 'catalog_time_zone_raw_offset_millis';
   static const tileExtentPrefixKey = 'tile_extent_';
   static const tileLayoutPrefixKey = 'tile_layout_';
   static const entryRenamingPatternKey = 'entry_renaming_pattern';
@@ -86,7 +86,7 @@ class SettingKeys {
   static const pinnedFiltersKey = 'pinned_filters';
   static const hiddenFiltersKey = 'hidden_filters';
   static const deactivatedHiddenFiltersKey = 'deactivated_hidden_filters';
-  static const showAlbumPickQueryKey = 'show_album_pick_query';
+  static const showTitleQueryPrefixKey = 'show_title_query_';
 
   // viewer
   static const viewerQuickActionsKey = 'viewer_quick_actions';